Adding a fuzzer for lossless jpeg decoding
The DNG file format is very complex. With the original fuzzer, it's very
hard to reach and explore the lossless jpeg decoding part of the code.
This fuzzer is designed to specifically targetting this code by calling
the decoding entrypoint with a binary blob directly. The seed file is
extracted from the original PoC.

Test: Manually run the fuzzer
Change-Id: I2d95fdb5708c84cc386e9b21ccf7ab9f06986854